Move system_restart_activation to orchestrator
